add RETRO_DEVICE_KEYBOARD support
By default, hotkeys are given precedence over keyboard input. Only if a
key press is not handled as part of an hotkey is it sent to the core as
keyboard input.

In addition, a game focus mode was introduced, managed by a toggle
hotkey that defaults to scroll lock. In this mode, all keyboard-bound
hotkeys are disabled, with the exception of the game focus toggle
itself, allowing most key presses to be provided to the core as keyboard
input.

Keyboard input is only supported by the poll-based retro_input_state_t
callback. The event-based retro_keyboard_callback is still not
supported.