add opengl hardware rendering support
If a core requests opengl hardware rendering, the video module creates a
framebuffer according to the core-provided requirements and sets the render
texture as its color attachment. The core gets the framebuffer through the
`retro_hw_get_current_framebuffer_t` libretro callback.

Every frame, after the core is done rendering, it calls `retro_video_refresh_t`
with `data = RETRO_HW_FRAME_BUFFER_VALID`, in which case we just make sure the 
view is set according to the provided `width` and `height` parameters, instead
of placing the contents of `data` into the texture as would happen in a software
rendering configuration.

When it's time for RALibretro to draw a frame, there is no difference between
software and hardware rendering: the texture is still rendered to a
full-window quad. This works because the core renders to the framebuffer to
which the rendered texture is set as a color attachment.

----

Hardware rendering cores can choose to use the opengl coordinate system where
(0,0) is bottom-left, whilst it would be top-left in the usual libretro coord
system that's used in software rendering.

If the core requests bottom-left coordinates, the vertex buffer is set so that
the texture is flipped vertically in 0º and 180º rotations, and flipped
horizontally in 90º and 270º rotations.

Images generated by screenshots and save states are also flipped vertically,
though in this case the flip is performed by the video module itself on the raw
texture data.

----

Because the core and the frontend share the opengl context, it's possible for
them to screw with each other by leaving non-default opengl state behind. To
protect the core from the frontend, the video module was updated to reset the
opengl state it uses. To protect the frontend from the core, the video
module sets a bunch of opengl state back to defaults after the core is done
rendering a frame (`postHwRenderReset`). This reset is also performed when we
switch from hardware to software rendering.

----

Finally, most of the code dealing with resizing the framebuffer, setting the
viewport, creating the vertex buffer, etc... was refactored into two methods:
`ensureFramebuffer` and `ensureView`. This avoids the duplication of the
somewhat complex logic that was getting hard to manage.

Related to this, the render texture's size is now only updated when the core
changes its max width or height, instead of whenever the frame's dimensions
(the view) change.