fix texture wrapping
The texture rendered to the whole-window quad was using `GL_REPEAT`
wrapping. Because coord 1.0 is slightly beyond the end of the texture,
the bottom row of pixels would be rendered as the result of filtering
the bottom and top rows of the texture.

Fixed by setting the texture's wrapping to `GL_CLAMP_TO_EDGE`.