wrap vertex buffer in vertex array
The vertex buffer was using globally bound state that was not being set
every frame. Cores, such as parallel_n64, could switch this global state
and cause the full-screen quad to not get rendered, for example.

Instead of setting the required global state every frame, the vertex
buffer is now wrapped by a vertex array that is capable of keeping track
of that state. Simply binding the vertex array is enough.