render frames inside the video refresh callback
RALibretro used to delay the rendering of frames until after the core
yielded back control from `retro_run`. This is ok for software rendering
cores, but hardware rendering cores are allowed to modify the
framebuffer and other OpenGL state between calling the video refresh
callback and returning from `retro_run`.

Libretro's reference frontend, RetroArch, avoids the issue by rendering
the frame inside the video refresh callback. The same solution is
implemented in this commit.

Some relevant implementation details:

* The `Video` component is now responsible for swapping the OpenGL
buffers and needs the SDL window handle to do so, which is managed by
`Application`. Access is granted through a `VideoContext` object that
provides a `swapBuffers` callback, passed in through `Video::init`.

* When frames were to be bypassed (due to turbo, for example),
`Application` used to simply not call `Video::draw`, which is no longer
possible. The replacement solution was to add a `generateVideo`
parameter to `Core::step`, similar to the existing `generateAudio`
parameter. If the parameter is `false`, the `Video` component will be
disabled and will not render a frame when the core calls the video
refresh callback.

* Since `Application` is no longer calling `Video::draw` on every
iteration of its main loop, it needs a wasy to explicitly clear the
screen when a game is not running. To do so it calls `Video::clear()`.
This is done on startup and when the current game is reset or unloaded.
Similarly, the `Video` module must explicitly redraw the frame when the
window is resized or the framebuffer is modified by loading a save
state.