unbind GL_ARRAY_BUFFER to avoid conflicts with the core's rendering
The `VecX` core's hardware rendering mode was not working. A single
frame was displayed before the screen turned black, even though the
game could be heard running in the background.

In `Video::ensureVertexArray`, a buffer is bound to the VAO upon
creation so that it doesn't have to be explicitly bound everytime we
draw. It seems, however, that even though the `GL_ARRAY_BUFFER` bind
affects the currently bound VAO, it also affects global state. To avoid
this conflicting with a core's own rendering, the buffer must be unbound
(after the VAO itself is unbound, to keep the association between the
two... not weird at all).

There was also a minor issue during the boot of the core, in which it
called the set-geometry callback with a max width and height of 0,
causing an error: `ensure framebuffer: failed to create hardware render
framebuffer 0 x 0`. The core did call set-geometry again soon after with
proper values, so the error wasn't noticeable in practice. Still, the
geometry's width and height values are now used in place of the
corresponding maximums, if the latter are smaller.